pyproject.toml
[tool.poetry]
name = "BKMS"
version = "0.1.0"
description = "Bibliography Knows Meaningful Suggestions"
authors = [
  "Jaeyeong Yang <jaeyeong.yang1125@gmail.com>"
]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
setuptools = "^62.3.2"
Flask = {extras = ["async"], version = "^2.1.2"}
python-dotenv = "^0.20.0"
neo4j = "5.0.0a1"
psycopg2-binary = "^2.9.3"
jupyterlab = "^3.4.1"
pandas = "^1.3.2"
tensorflow = { version = "^2.9.0", markers = "sys_platform == 'linux'" }
tensorflow-hub = { version = "^0.12.0", markers = "sys_platform == 'linux'" }

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
